requirements.txt
annotated-types==0.6.0
exceptiongroup==1.2.0
iniconfig==2.0.0
lxml==5.1.0
packaging==23.2
pluggy==1.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
python-docx==1.1.0
tomli==2.0.1
typing_extensions==4.9.0
